Move DBConnection to new files...
... And SqliteSampleBlock does not depend on ProjectFileIO.
